Replaced Access Control component with AdminLevelAC + moved DebugComponent to separate file
